fix: Storage for the page is immutable when saving page with kirby-seo fields
Trying the fix the `Storage for the page is immutable` error message that’s popping when using the plugin with Kirby 5.x.

This basically only implements what Nico suggested in the Discord chat. 

Currently i can’t reproduce the issue reliably so i don’t know if this will fix it. Maybe someone can give it a try or let me know how to reproduce?

Signed-off-by: Ove Numrich <47149622+ovenum@users.noreply.github.com>